+++ CHANGELOG.md
@@ -25,3 +25,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `HttpRequestUtil.request` documentation
 - Update license year
 - Add `X-Request-Id` header
+- Extract `User-Agent` header from package.json
+- Rename `HTTPResponseError` -> `HttpResponseError`